formalized rebinding code
