+++ Assets/Scripts/Misc/Utils.cs
@@ -1,5 +1,8 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
 using UnityEngine;
+using UnityEngine.InputSystem;
 
 public static class Utils
 {
@@ -19,5 +22,21 @@ public static class Utils
     {
         if (obj.TryGetComponent(out result)) return;
         else throw new Exception($"`{obj.name}` is missing component of type `{typeof(T).Name}`!");
+    }
+
+    /// <summary>
+    /// Gets the action the player input has for a given action id.
+    /// </summary>
+    /// <param name="input">The player input in question</param>
+    /// <param name="actionId">The ID of the action</param>
+    /// <returns>The player input's copy of the action</returns>
+    public static InputAction PlayerInputActionOfActionId(this PlayerInput input, Guid actionId) {
+        var actionIter = input.actions.actionMaps.Aggregate(
+            new List<InputAction>().AsEnumerable(),
+            (aux, next) => aux.Concat(next).Concat(next)
+        );
+        return actionIter
+            .Where(action => action.id == actionId)
+            .First();
     }
 }
